@@ -14,6 +14,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using ServiceLayer.DatabaseServices.Concrete;
 
@@ -50,6 +51,8 @@ namespace EfCoreInAction
             }
             services.AddDbContext<EfCoreContext>(options => options.UseSqlServer(connection,
                 b => b.MigrationsAssembly("DataLayer")));
+            services.AddControllersWithViews();
+            services.AddRazorPages();
 
             //Add AutoFac
             var containerBuilder = new ContainerBuilder();
@@ -60,7 +63,7 @@ namespace EfCoreInAction
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, 
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, 
             ILoggerFactory loggerFactory, IHttpContextAccessor httpContextAccessor)
         {
             //Remove the standard loggers because they slow the applictaion down
@@ -70,20 +73,24 @@ namespace EfCoreInAction
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
-                app.UseBrowserLink();
             }
             else
             {
                 app.UseExceptionHandler("/Home/Error");
+                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+                app.UseHsts();
             }
+            app.UseHttpsRedirection();
+            app.UseStaticFiles();
 
             app.UseStaticFiles();
 
-            app.UseMvc(routes =>
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapRoute(
+                endpoints.MapControllerRoute(
                     name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
+                    pattern: "{controller=Home}/{action=Index}/{id?}");
+                endpoints.MapRazorPages();
             });
         }
     }
