@@ -38,10 +38,11 @@ namespace test.UnitTests.DataLayer
         public void TestDeleteBookInLineItemFails()
         {
             //SETUP
-            var inMemDb = new SqliteInMemory();
+            var options = SqliteInMemory.CreateOptions<EfCoreContext>();
 
-            using (var context = inMemDb.GetContextWithSetup())
+            using (var context = new EfCoreContext(options))
             {
+                context.Database.EnsureCreated();
                 context.SeedDatabaseFourBooks();
                 var userId = Guid.NewGuid();
 
@@ -61,13 +62,15 @@ namespace test.UnitTests.DataLayer
                 };
                 context.Orders.Add(order);
                 context.SaveChanges();
-
+            }
+            using (var context = new EfCoreContext(options))
+            {
                 //ATTEMPT
                 context.Books.Remove(context.Books.First());
-                var ex = Assert.Throws<InvalidOperationException>(() => context.SaveChanges());
+                var ex = Assert.Throws<DbUpdateException>(() => context.SaveChanges());
 
                 //VERIFY
-                ex.Message.ShouldEqual("The association between entity types 'Book' and 'LineItem' has been severed but the foreign key for this relationship cannot be set to null. If the dependent entity should be deleted, then setup the relationship to use cascade deletes.");
+                ex.InnerException.Message.ShouldEqual("SQLite Error 19: 'FOREIGN KEY constraint failed'.");
             }
         }
 
